# docs/openapi.yaml
openapi: 3.0.3
info:
  title: ERPGo API
  description: |
    A comprehensive Enterprise Resource Planning (ERP) system built with Go, providing user management, inventory management, order management, and business analytics capabilities.

    ## Overview
    ERPGo is a modern, cloud-native ERP system designed for high performance, scalability, and maintainability. It provides:

    - User authentication and authorization with JWT tokens
    - Product catalog management with categories and variants
    - Inventory tracking and warehouse management
    - Order processing and fulfillment workflows
    - Customer management and address handling
    - Comprehensive analytics and reporting

    ## Base URLs
    - **Production**: `https://api.erpgo.example.com`
    - **Staging**: `https://api-staging.erpgo.example.com`
    - **Development**: `http://localhost:8080`

    ## Authentication
    The API uses JWT (JSON Web Tokens) for authentication. Include the token in the `Authorization` header:

    ```
    Authorization: Bearer <your-jwt-token>
    ```

    ## Rate Limiting
    API requests are rate-limited to ensure fair usage:
    - **Default**: 100 requests per minute
    - **Burst**: 200 requests
    - **Window**: 1 minute

    Rate limit headers are included in responses:
    ```
    X-RateLimit-Limit: 100
    X-RateLimit-Remaining: 99
    X-RateLimit-Reset: 1640995200
    ```
  version: 1.0.0
  contact:
    name: ERPGo Support
    email: support@erpgo.example.com
...
  license:
...
servers:
...
paths:
...
  /api/v1/users:
    get:
      tags:
        - Users
      summary: List users
      description: Get paginated list of users with filtering
      operationId: listUsers
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Search'
        - name: is_active
          in: query
          schema:
            type: boolean
          description: Filter by active status
        - name: is_verified
          in: query
          schema:
            type: boolean
          description: Filter by verification status
        - name: role_id
          in: query
          schema:
            type: string
            format: uuid
          description: Filter by role ID
        - $ref: '#/components/parameters/SortBy'
        - $ref: '#/components/parameters/SortOrder'
      responses:
        '200':
          description: Users retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListUsersResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/users/{id}:
    get:
      tags:
        - Users
      summary: Get user by ID
      description: Retrieve user information by ID
      operationId: getUser
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        '200':
          description: User retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserInfo'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags:
        - Users
      summary: Update user
      description: Update user information
      operationId: updateUser
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/UserId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: User updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserInfo'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'

    delete:
      tags:
        - Users
      summary: Delete user
      description: Delete user account (soft delete)
      operationId: deleteUser
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        '204':
          description: User deleted successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # Product management endpoints
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
      description: Page number for pagination
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: Number of items per page
    Search:
      name: search
      in: query
      schema:
        type: string
        minLength: 1
      description: Search term
    SortBy:
      name: sort_by
      in: query
      schema:
        type: string
        enum: [created_at, updated_at, name, email, price]
      description: Field to sort by
    SortOrder:
      name: sort_order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: desc
      description: Sort order
    UserId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: User ID
...
  schemas:
...
    UserInfo:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        email:
          type: string
          format: email
          example: "user@example.com"
        username:
          type: string
          example: "username"
        first_name:
          type: string
          example: "John"
        last_name:
          type: string
          example: "Doe"
        phone:
          type: string
          example: "+1234567890"
        roles:
          type: array
          items:
            type: string
          example: ["user", "admin"]
        is_active:
          type: boolean
          example: true
        is_verified:
          type: boolean
          example: true
        last_login_at:
          type: string
          format: date-time
          example: "2024-01-08T12:00:00Z"
        created_at:
          type: string
          format: date-time
          example: "2024-01-08T10:30:00Z"
    UpdateUserRequest:
      type: object
      properties:
        first_name:
          type: string
          example: "John"
        last_name:
          type: string
          example: "Doe"
        phone:
          type: string
          example: "+1234567890"
        is_active:
          type: boolean
          example: true
    ListUsersResponse:
      type: object
      properties:
        users:
          type: array
          items:
            $ref: '#/components/schemas/UserInfo'
        pagination:
          $ref: '#/components/schemas/Pagination'

    # Product schemas
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 20
        total:
          type: integer
          example: 150
        total_pages:
          type: integer
          example: 8
        has_next:
          type: boolean
          example: true
        has_prev:
          type: boolean
          example: false
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              example: "VALIDATION_ERROR"
            message:
              type: string
              example: "Request validation failed"
            details:
              oneOf:
                - type: array
                  items:
                    type: object
                    properties:
                      field:
                        type: string
                        example: "email"
                      message:
                        type: string
                        example: "Invalid email format"
                - type: string
                  example: "Additional error details"
            timestamp:
              type: string
              format: date-time
              example: "2024-01-08T10:30:00Z"
            request_id:
              type: string
              example: "req_123456789"
  responses:
    BadRequest:
      description: Bad request - Invalid request data
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            error:
              code: "BAD_REQUEST"
              message: "Invalid request data"
              timestamp: "2024-01-08T10:30:00Z"
              request_id: "req_123456789"
    Unauthorized:
      description: Unauthorized - Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            error:
              code: "UNAUTHORIZED"
              message: "Authentication required"
              timestamp: "2024-01-08T10:30:00Z"
              request_id: "req_123456789"
    Forbidden:
      description: Forbidden - Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            error:
              code: "FORBIDDEN"
              message: "Insufficient permissions"
              timestamp: "2024-01-08T10:30:00Z"
              request_id: "req_123456789"
    NotFound:
      description: Not found - Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            error:
              code: "NOT_FOUND"
              message: "Resource not found"
              timestamp: "2024-01-08T10:30:00Z"
              request_id: "req_123456789"
...
    ValidationError:
      description: Validation error - Request validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            error:
              code: "VALIDATION_ERROR"
              message: "Request validation failed"
              details:
                - field: "email"
                  message: "Invalid email format"
              timestamp: "2024-01-08T10:30:00Z"
              request_id: "req_123456789"
...